worker/requirements.txt
pika
psycopg2-binary
mysql-connector-python